backend/requirements.txt
delocate==0.10.4
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
pytesseract @ git+https://github.com/madmaze/pytesseract.git@251c572d1e04e90c826116c8412a6ea04a011578
tesseract==0.1.3
typing_extensions==4.5.0
